pyproject.toml
[project]
name = "tdah-printer"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.47",
    "pillow>=10.0",
    "pyusb>=1.2.1",
    "python-escpos>=3.0a9",
]
